Moved processing folder of SVG UI to menu item.
The button in the document window UI did not make any sense for processing a folder of svg files. This is now done from a menu item.
